% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_genes_7.R
\name{cell_genes_7}
\alias{cell_genes_7}
\title{Cell genes}
\usage{
cell_genes_7(
  sce,
  group = "annotation",
  group_cell_1 = "B_cell",
  group_cell_2 = "NK_cell",
  Two_cell_enrichment = TRUE,
  dir = "./6.Cell genes"
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{group_cell_1}{Select cell type 1.}

\item{group_cell_2}{Select cell type 2.}

\item{Two_cell_enrichment}{Analysis of differential gene enrichment in two types of cells.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Selection and enrichment analysis of differential genes in cells.
}
