% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_markers_6.R
\name{cell_markers_6}
\alias{cell_markers_6}
\title{Cell markers}
\usage{
cell_markers_6(
  sce,
  group = "seurat_clusters",
  method = c("SingleR", "personalization"),
  cells_markers = c("PTPRC", "CD3D", "CD3E", "CD4", "CD8A", "CD19", "CD79A", "MS4A1",
    "IGHG1", "MZB1", "SDC1", "CD68", "CD163", "CD14", "LAMP3", "IDO1", "IDO2", "CD1E",
    "CD1C", "KLRB1", "NCR1", "FGF7", "MME", "ACTA2", "PECAM1", "VWF"),
  dir = "./5.Cell markers"
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{method}{Cell annotation method.If personalization is selected, run cell_markers_6_personalization.}

\item{cells_markers}{Cell marker genes were selected.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Cellular annotation of a single cell subpopulation.
}
