% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_markers_6_personalization.R
\name{cell_markers_6_personalization}
\alias{cell_markers_6_personalization}
\title{Cell markers personalization}
\usage{
cell_markers_6_personalization(
  sce,
  group = "seurat_clusters",
  ClusterID = c("0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11"),
  celltype = c("T cells", "T cells", "B cells", "T cells", "Monocytes and macrophages",
    "Monocytes and macrophages", "T cells", "Monocytes and macrophages", "T cells",
    "T cells", "T cells", "Monocytes and macrophages"),
  dir = "./5.Cell markers"
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{ClusterID}{Number of cell subsets.The parameter must correspond to cellType.}

\item{celltype}{Number of celltype.The parameter must correspond to ClusterID.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Manual annotation of single cell subsets.
}
