% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSce.R
\name{compare_sce}
\alias{compare_sce}
\title{Use SCPA to compare pathways within a SingleCellExperiment object}
\usage{
compare_sce(
  sce_object,
  assay_name = "logcounts",
  group1 = NULL,
  group1_population = NULL,
  group2 = NULL,
  group2_population = NULL,
  pathways,
  downsample = 500
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object with populations defined in the column data}

\item{assay_name}{Assay name to extract expression values from. Defaults to logcounts}

\item{group1}{First comparison group as defined by \code{colData()} columns
of SingleCellExperiment object e.g. cell_type}

\item{group1_population}{Populations within group1 to compare
e.g. c(t_cell, b_cell)}

\item{group2}{Second comparison group as defined by \code{colData()} columns
of SingleCellExperiment object e.g. hour}

\item{group2_population}{Population within group2 to compare
e.g. 24}

\item{pathways}{Pathway gene sets with each pathway in a separate list. For formatting of
gene lists, see documentation at https://jackbibby1.github.io/SCPA/articles/using_gene_sets.html}

\item{downsample}{Option to downsample cell numbers. Defaults to 500 cells per condition. If a population
has < 500 cells, all cells from that condition are used.}
}
\value{
Statistical results from the SCPA analysis. The qval should be the
primary metric that is used to interpret pathway differences i.e. a higher
qval translates to larger pathway differences between conditions.
If only two samples are provided, a fold change (FC) enrichment score will also be
calculated. The FC output is generated from a running sum of mean changes in gene
expression from all genes of the pathway. It's calculated from average pathway
expression in population1 - population2, so a negative FC means the pathway is
higher in population2.
}
\description{
This function takes a SingleCellExperiment object as an input, and
compares gene sets over specified conditions/populations.
}
\examples{
\dontrun{
scpa_out <- compare_sce(
     group1 = "cell",
     group1_population = c("t_cell", "b_cell"),
     group2 = "hour",
     group2_population = c("24"),
     pathways = pathways)
}

}
