% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSeurat.R
\name{compare_seurat}
\alias{compare_seurat}
\title{Use SCPA to compare pathways within a Seurat object}
\usage{
compare_seurat(
  seurat_object,
  group1 = NULL,
  group1_population = NULL,
  group2 = NULL,
  group2_population = NULL,
  pathways,
  downsample = 500
)
}
\arguments{
\item{seurat_object}{Seurat object with populations defined in the meta data}

\item{group1}{First comparison group as defined by meta data in
Seurat object e.g. cell_type}

\item{group1_population}{Population within group1 to compare
e.g. t_cell}

\item{group2}{Second comparison group as defined by column names in
Seurat object e.g. hour}

\item{group2_population}{Population within group2 to compare
e.g. 24}

\item{pathways}{Pathway gene sets with each pathway in a separate list. For formatting of
gene lists, see documentation at https://jackbibby1.github.io/SCPA/articles/using_gene_sets.html}

\item{downsample}{Option to downsample cell numbers. Defaults to 500 cells per condition. If a population
has < 500 cells, all cells from that condition are used.}
}
\value{
Statistical results from the multicross tool.
If only two samples are provided, an enrichment score will also be
calculated.
}
\description{
This function takes a Seurat object as an input, and
compares gene sets over specified conditions/populations.
}
\examples{
\dontrun{
scpa_out <- compare_seurat(
     list(sample1, sample2, sample3),
     pathways = pathways)
}

}
