% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_2.R
\name{data_quality_2}
\alias{data_quality_2}
\title{Data quality}
\usage{
data_quality_2(sce, group = "orig.ident", dir = "./1.Data quality")
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Summary of single cell data quality.
}
