% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers_5.R
\name{find_markers_5}
\alias{find_markers_5}
\title{Find markers}
\usage{
find_markers_5(
  sce,
  group = "seurat_clusters",
  genes_list = c("IL7R", "CD3E", "CD8A", "CD3D", "CD2"),
  dir = "./4.Find markers"
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{genes_list}{The expression value of selected genes was visualized.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Personalized gene visualization and differential gene selection.
}
