% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatPathways.R
\name{format_pathways}
\alias{format_pathways}
\title{Format the output of msigdbr to use with SCPA}
\usage{
format_pathways(msigdbr_output)
}
\arguments{
\item{msigdbr_output}{output from msigdbr}
}
\value{
list of pathways with corresponding genes
}
\description{
This function takes the output of msigdbr
and formats the pathways into something that
can be used within SCPA
}
\examples{
\dontrun{
pathways <- format_pathways(
     pathways
)
}

}
