% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_objects_2.R
\name{integrate_objects_2}
\alias{integrate_objects_2}
\title{Data integration}
\usage{
integrate_objects_2(
  sce,
  group = "orig.ident",
  input.type = c("list", "merged"),
  variablefeature.method = "vst",
  normalization.method.selection = c("LogNormalize", "SCT"),
  reduction.selection = c("cca", "rpca", "rlsi")
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{input.type}{Single cell input type data}

\item{variablefeature.method}{Variablefeature filtering method}

\item{normalization.method.selection}{Standardized methods for single cell data}

\item{reduction.selection}{Single cell data integration method}
}
\value{
a sce object
}
\description{
Multiple single cell data were combined and integrated.
}
