% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMatrices.R
\name{pathway_matrices}
\alias{pathway_matrices}
\title{Create pathway matrices from gene sets}
\usage{
pathway_matrices(samples, pathways, sample_names = NULL)
}
\arguments{
\item{samples}{Expression matrix for each population, with each population
as a separate list. These matrices can be generated with e.g. \code{SCPA::seurat_extract()}}

\item{pathways}{Either a gene set gmt file, csv file, or output from \code{SCPA::format_pathways()}
that contains your gene sets of interest. Information on formatting gene sets can be found
at https://jackbibby1.github.io/SCPA/articles/using_gene_sets.html.}

\item{sample_names}{Names of the samples to be used in the output}
}
\description{
This function takes a matrix and pathway csv/gmt files
and input and creates expression matrices for each
pathway. The resulting output is a nested list of expression
matrices for each pathway by each sample.
}
\examples{
\dontrun{
pathway_expression <- pathway_matrices(
     list(sample1, sample2),
     pathways = "your/path_to_gmt",
     sample_names = c("population1", "population2"))
}

}
