% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHeatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap of qvals from SCPA}
\usage{
plot_heatmap(
  scpa_out,
  highlight_pathways = NULL,
  row_fontsize = 6,
  column_fontsize = 10,
  column_names = colnames(scpa_out),
  show_row_names = FALSE,
  cluster_columns = TRUE,
  hm_colors = NULL,
  scale_breaks = NULL
)
}
\arguments{
\item{scpa_out}{Data frame containing Pathways and qvals generated from
compare_pathways}

\item{highlight_pathways}{Chosen pathway or pathways to highlight on the heatmap,
supplied as character vector. If no argument is given, all pathways are
shown}

\item{row_fontsize}{Font size of pathway names}

\item{column_fontsize}{Font size of the sample names}

\item{column_names}{Option to supply names of the heatmap
columns if more than one populations is present}

\item{cluster_columns}{Should columns in the heatmap be clustered?}

\item{hm_colors}{Colors to be used in the heatmap}

\item{scale_breaks}{Breaks to use in the colors of the heatmap.
Length must be equal to the number of colors.}
}
\value{
list of pathways with corresponding genes
}
\description{
This function takes the output from SCPA and plots a
heatmap of the qvals
}
\examples{
\dontrun{
plot_heatmap(
     scpa_out = scpa_result,
     pathway = "mtorc"
)
}

}
