% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRank.R
\name{plot_rank}
\alias{plot_rank}
\title{Plot the rank of specific pathways from SCPA output}
\usage{
plot_rank(
  scpa_out,
  pathway,
  population_name = "qval",
  base_point_size = 2,
  base_point_color = "royalblue2",
  highlight_point_size = 3,
  highlight_point_color = "orangered2",
  label_pathway = T,
  label_size = 4
)
}
\arguments{
\item{scpa_out}{Data frame containing Pathways and qvals generated from
compare_pathways}

\item{pathway}{Chosen pathway or pathways to plot the rank of. This can
be specific e.g. HALLMARK_GLYCOLYSIS to plot a specific result, or
generic e.g. glycolysis to plot all glycolysis pathways}

\item{population_name}{Column name of the population to plot, if more
than one population present in the data frame}

\item{base_point_size}{Size of the base points to plot on the graph}

\item{base_point_color}{Color of the base points to plot on the graph}

\item{highlight_point_size}{Size of the highlighted points to plot on the graph}

\item{highlight_point_color}{Color of the highlighted points to plot on the graph}

\item{label_pathway}{Should the selected pathway be labelled?}

\item{label_size}{Text size of the pathway label}
}
\value{
list of pathways with corresponding genes
}
\description{
This function takes the output from SCPA and plots the
rank of a user defined pathway.
}
\examples{
\dontrun{
plot_rank(
     scpa_out = scpa_result,
     pathway = "interferon",
     population_name = cd4_qval,
)
}

}
