% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_1.R
\name{prepare_1}
\alias{prepare_1}
\title{Preparation}
\usage{
prepare_1(run)
}
\arguments{
\item{run}{Define operation parameters}
}
\value{
a character
}
\description{
Single cell data R package reference and environment construction.
}
