% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_3.R
\name{quality_control_3}
\alias{quality_control_3}
\title{Data filtration}
\usage{
quality_control_3(
  sce,
  group = "orig.ident",
  nFeature_lower = 300,
  nFeature_upper = 8000,
  nCount_lower = 3,
  nCount_upper = 10000,
  percent_mito_lower = 0,
  percent_mito_upper = 25,
  percent_hb_upper = 10,
  percent_hb_lower = 0,
  percent_ribo_lower = 3,
  percent_ribo_upper = 1000,
  dir = "./2.Quality control"
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{nFeature_lower}{The minimum number of genes detected per cell.}

\item{nFeature_upper}{The maximum number of genes detected per cell.}

\item{nCount_lower}{The minimum number of UMI per cell.}

\item{nCount_upper}{The maximum number of UMI per cell.}

\item{percent_mito_lower}{The minimum proportion of mitochondria.}

\item{percent_mito_upper}{The maximum proportion of mitochondria.}

\item{percent_hb_upper}{The maximum proportion of hemoglobin genes detected.}

\item{percent_hb_lower}{The minimum proportion of hemoglobin genes detected.}

\item{percent_ribo_lower}{The minimum gene expression ratio of ribosomal protein.}

\item{percent_ribo_upper}{The maximum gene expression ratio of ribosomal protein.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Quality control and data screening for single cell data.
}
