% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomCells.R
\name{random_cells}
\alias{random_cells}
\title{Randomly sample cells (columns)}
\usage{
random_cells(df, n)
}
\arguments{
\item{df}{Data frame or matrix.}

\item{n}{Number of cells to sample.}
}
\value{
Matrix or data frame of n randomly sampled rows
}
\description{
This function takes a matrix or data frame as
input and randomly samples columns
}
\examples{
\dontrun{
df <- matrix(1:1000, 500, 20)
sub_df <- random_cells(
    df = df,
    n = 20
)
}
}
