% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceExtract.R
\name{sce_extract}
\alias{sce_extract}
\title{Extract data from a SingleCellExperiment object}
\usage{
sce_extract(
  sce_object,
  assay_name = "logcounts",
  meta1 = NULL,
  value_meta1 = NULL,
  meta2 = NULL,
  value_meta2 = NULL,
  pseudocount = 0.001
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object containing expression data}

\item{assay_name}{Name of assay to pull from. Defaults to "logcounts"}

\item{meta1}{Metadata name to subset}

\item{value_meta1}{Value to select within \code{meta1} function}

\item{meta2}{Metadata name to subset}

\item{value_meta2}{Value to select within \code{meta2} function}

\item{pseudocount}{Pseudocount to add to data. Defaults to 0.001}
}
\value{
Matrix containing count values of selected populations
}
\description{
This function takes a SingleCellExperiment object as an input, and returns
a matrix based on subsetting parameters. Either none, one, or
two metadata features can be selected for a given input.
}
\examples{
\dontrun{
cd4 <- sce_extract(
   sce_object,
   meta1 = "Hour",
   value_meta1 = 12,
   meta2 = "Cell_Type",
   value_meta2 = "CD4"
)
}

}
