% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratExtract.R
\name{seurat_extract}
\alias{seurat_extract}
\title{Extract Data From A Seurat Object}
\usage{
seurat_extract(
  seu_obj,
  assay = "RNA",
  meta1 = NULL,
  value_meta1 = NULL,
  meta2 = NULL,
  value_meta2 = NULL,
  pseudocount = 0.001
)
}
\arguments{
\item{seu_obj}{Seurat object}

\item{assay}{Seurat assay to extract. Defaults to RNA}

\item{meta1}{Metadata column to subset}

\item{value_meta1}{Value to select within \code{meta1} column}

\item{meta2}{Metadata column to subset}

\item{value_meta2}{Value to select within \code{meta2} column}

\item{pseudocount}{Pseudocount to add to data. Defaults to 0.001}
}
\value{
Matrix containing count values of selected populations
}
\description{
This function takes a Seurat object as an input, and returns
an expression matrix based on subsetting parameters. Either none, one, or
two metadata features can be selected for a given input.
}
\examples{
\dontrun{
cd4 <- seurat_extract(
   Seurat_Object,
   meta1 = "Hour",
   value_meta1 = 12,
   meta2 = "Cell_Type",
   value_meta2 = "CD4"
)
}

}
