% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_flow_4.R
\name{standard_flow_4}
\alias{standard_flow_4}
\title{Standard flow}
\usage{
standard_flow_4(
  sce,
  group = "orig.ident",
  dim_number = 20,
  dbs_rate = NULL,
  remove_doublets = TRUE,
  nmf_rank_number = 10,
  Scale_cellcycle = FALSE,
  resolution_low = 0.1,
  resolution_up = 1,
  resolution_by = 0.1,
  dir = "./3.scRNAflow"
)
}
\arguments{
\item{sce}{Single cell Seurat data.}

\item{group}{Grouping data of single cells.}

\item{dim_number}{The number of dimension reduction.}

\item{dbs_rate}{The expected doublet rate.}

\item{remove_doublets}{Whether to remove doublet data.}

\item{nmf_rank_number}{Choice of the factorization rank.}

\item{Scale_cellcycle}{Remove cell cycle effects.}

\item{resolution_low}{The minimum grouping parameter.}

\item{resolution_up}{The maximum grouping parameter.}

\item{resolution_by}{The interval of grouping parameters.}

\item{dir}{Output data directory.}
}
\value{
a sce object
}
\description{
Dimension reduction, clustering and clustering of single cell data were carried out.
}
