% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePathways.R
\name{compare_pathways}
\alias{compare_pathways}
\title{Use SCPA to compare gene sets}
\usage{
compare_pathways(
  samples,
  pathways,
  downsample = 500,
  min_genes = 15,
  max_genes = 500
)
}
\arguments{
\item{samples}{List of samples, each supplied as an expression matrix}

\item{pathways}{List of pathways and their genes}

\item{downsample}{Option to downsample cell numbers. Default is 500}

\item{min_genes}{Gene sets with fewer than this number of genes will be excluded}

\item{max_genes}{Gene sets with more than this number of genes will be excluded}
}
\value{
Statistical results from the SCPA analysis.
If only two samples are provided, an enrichment score will also be
calculated.
}
\description{
This function takes an input of samples and pathways
to compare gene sets over different conditions.
}
\examples{
\dontrun{
scpa_result <- compare_pathways(
     list(sample1, sample2, sample3),
     pathways = pathways)
}

}
