% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPaths.R
\name{get_paths}
\alias{get_paths}
\title{Convert gene sets to filtered pathway list}
\usage{
get_paths(pathway_filepath, min_genes = 15, max_genes = 500)
}
\arguments{
\item{pathway_filepath}{filepath to csv file, or list object}

\item{min_genes}{Minimum number of genes required in a pathway for inclusion}

\item{max_genes}{Maximum number of genes required in a pathway for inclusion}
}
\value{
list of pathways with corresponding genes
}
\description{
This function takes one of two inputs:
\enumerate{
\item csv file where pathway title is in the first column
and subsequent columns contain genes of that pathway
\item list where pathway name is contained in the first column
and the genes of that pathway are in the second column
}
}
\examples{
\dontrun{
pathways <- get_paths(
     "Documents/gene_list.csv"
)
}

}
