% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMatrices.R
\name{pathway_matrices}
\alias{pathway_matrices}
\title{Create pathway matrices from gene sets}
\usage{
pathway_matrices(samples, pathways, sample_names = NULL, pathway_names = NULL)
}
\arguments{
\item{samples}{Expression file for each population, with each population
as a separate list}

\item{pathways}{List of pathways to create pathway specific matrices.}

\item{sample_names}{Names of the samples to be used in output}

\item{pathway_names}{Names of pathways to be used in output}
}
\description{
This function takes a matrix and pathway list as
and input and creates expression matrices for each
pathway. The resulting output is a list of expression
matrices for each pathway
}
